Refactor type checker.
