+++ jl4/src/L4/Lexer.hs
@@ -874,3 +874,29 @@ posTokenCategory =
     TLineComment _ -> CComment
     TBlockComment _ -> CComment
     EOF -> CEOF
+
+-- | We ignore the file name, because we assume this has already been checked.
+inRange :: SrcPos -> SrcRange -> Bool
+inRange (MkSrcPos l c) (MkSrcRange (MkSrcPos l1 c1) (MkSrcPos l2 c2) _) =
+     (l, c) >= (l1, c1)
+  && (l, c) <= (l2, c2)
+
+-- TODO: eventually, we may want to take sufficient info so that we can print the file path
+-- for "external" ranges, but not ranges in the current file.
+prettySrcRange :: Maybe FilePath -> Maybe SrcRange -> Text
+prettySrcRange fp Nothing = prettyFilePath fp <> "<unknown range>"
+prettySrcRange fp (Just (MkSrcRange p1 p2 _)) = prettyFilePath fp <> prettySrcPos p1 <> prettyPartialSrcPos p1 p2
+
+prettyFilePath :: Maybe FilePath -> Text
+prettyFilePath Nothing   = ""
+prettyFilePath (Just fp) = Text.pack fp <> ":"
+
+prettySrcPos :: SrcPos -> Text
+prettySrcPos (MkSrcPos l c) = Text.show l <> ":" <> Text.show c
+
+prettyPartialSrcPos :: SrcPos -> SrcPos -> Text
+prettyPartialSrcPos (MkSrcPos rl rc) p@(MkSrcPos l c)
+  | rl == l && rc == c = ""
+  | rl == l            = "-" <> Text.show c
+  | otherwise          = "-" <> prettySrcPos p
+
